refactor: move move_player_by_table system to willo module
